fix : toggle 상태 적용 #263
