+++ front/src/component/Modal/SetUser.tsx
@@ -9,6 +9,7 @@ import { axiosInstance } from '@/utils/axios';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 import { invalidMsgState, modalState } from '@/utils/recoil/atom';
 import InvalidMsg from './InvalidMsg';
+import SetUserToggle from '../Toggle/setUserToggle';
 
 interface SetUserProps {
   userid: string;
@@ -20,23 +21,9 @@ export default function SetUser({ userid, nickname, channelid }: SetUserProps) {
   const data = useSwrFetcher<IUserSetting>(
     `/channel/setting/${channelid}/${userid}`
   );
-
-  const setModal = useSetRecoilState(modalState);
-  const [admin, onChangeAdmin] = useToggle(data?.admin ?? false);
-  const [mute, onChangeMute] = useToggle(data?.mute ?? false);
-  const [ban, onChangeBan] = useToggle(data?.mute ?? false);
   const [invalidMsg] = useRecoilState(invalidMsgState);
+  const setModal = useSetRecoilState(modalState);
 
-  const banUserHandler = async () => {
-    axiosInstance
-      .put(`/channel/setting/${channelid}/user`, {
-        id: userid,
-        status: 'ban',
-      })
-      .then(() => {
-        onChangeBan();
-      });
-  };
   const kickUserHandler = async () => {
     axiosInstance
       .put(`/channel/setting/${channelid}/user`, {
@@ -47,28 +34,8 @@ export default function SetUser({ userid, nickname, channelid }: SetUserProps) {
         setModal(null);
       });
   };
-  const adminUserHandler = async () => {
-    axiosInstance
-      .put(`/channel/setting/${channelid}/user`, {
-        id: userid,
-        status: 'admin',
-      })
-      .then(() => {
-        onChangeAdmin();
-      });
-  };
 
-  const muteUserHandler = async () => {
-    axiosInstance
-      .put(`/channel/setting/${channelid}/user`, {
-        id: userid,
-        status: 'mute',
-      })
-      .then(() => {
-        onChangeMute();
-      });
-  };
-
+  if (!data) return;
   return (
     <Wrapper>
       <WrapperSection>
@@ -77,24 +44,7 @@ export default function SetUser({ userid, nickname, channelid }: SetUserProps) {
           <div>{nickname}</div>
         </div>
         <div>
-          <Toggle
-            text='admin'
-            color='green'
-            checked={admin}
-            onToggle={adminUserHandler}
-          />
-          <Toggle
-            text='mute'
-            color='green'
-            checked={mute}
-            onToggle={muteUserHandler}
-          />
-          <Toggle
-            text='ban'
-            color='green'
-            checked={ban}
-            onToggle={banUserHandler}
-          />
+          <SetUserToggle data={data} channelid={channelid} userid={userid} />
         </div>
       </WrapperSection>
       <InvalidMsg text={invalidMsg} />
